Re-added button to start guild join discord flow
